work on matdia

+++ src/app/shared/components/products/product/product.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit, inject } from '@angular/core';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Iproduct } from 'src/app/shared/models/products.interface';
 import { ProductsService } from 'src/app/shared/services/products.service';
+import { ConfirmDialogComponent } from '../../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-product',
@@ -16,7 +18,9 @@ export class ProductComponent implements OnInit {
   private _route = inject(ActivatedRoute);
   private _router = inject(Router);
   private _productsService = inject(ProductsService);
-  constructor() { }
+  constructor(
+    private _matDialog : MatDialog
+  ) { }
 
   ngOnInit(): void {
     // console.log(this._route.snapshot.params);
@@ -36,8 +40,23 @@ export class ProductComponent implements OnInit {
 
 
   onRemoveProduct(){
-    this._productsService.removeProduct(this.productObj);
-    this._router.navigate(['/products'])
+
+    let dialogConf =  new MatDialogConfig();
+    dialogConf.width = "350px";
+    dialogConf.data = `Are you sure, you want to remove product ${this.productObj.pname}`
+    
+    let dialogRef = this._matDialog.open(ConfirmDialogComponent, dialogConf)
+
+
+    dialogRef.afterClosed()
+    .subscribe(res => {
+      if(res){
+        this._productsService.removeProduct(this.productObj);
+        this._router.navigate(['/products'])
+      }else{
+        return
+      }
+    })
   }
 
 }
